Created a Client Object
